// gestaoemprestimosfront-master/src/models/consultas.dto.ts
export class ConsultasDTO {
    idConsultas: number;
    nomeConsulta: string;
    caminhoArquivo: string;
    dataConsultas: string;
    listaCpf?: string;
    statusConsulta: string;
    downloadConsulta: string;
    quantConsulta: number;
    porcentagemConsulta: number;
    tipoConsulta: string;
}